update-submodule: Make target handling more robust.
Previously we would always just use origin for the remote,
add possibility to define the target more precisely.


For example --target something/foo
            --target other_remote:tag_name
            --target foo

If something/foo notation is used, then anything before / is used as
the remote.

If other_remote:tag_name is used, then other_remote is used as remote.

If only foo is used, then origin is used as the remote by default.